refactor : add a `ui` directory to the `screen` layer where it is missing
