--- src/screen/user/index.ts
@@ -1,1 +1,0 @@
-export { default as UserScreen } from './UserScreen'
