+++ src/screen/user/ui/index.ts
@@ -1,0 +1,1 @@
+export { default as UserScreen } from './UserScreen';
